+++ manager_server.py
@@ -74,6 +74,10 @@ class HeadlessClientInstance(RemoteHeadlessClient):
                 sleep(.5)
             # TODO: Purge old sessions.
 
+    # TODO: Check whether these are completely empty. This should only occur
+    # during the brief period of time between the headless client being ready
+    # and these being polled for the first time.
+
     def worlds(self):
         if not self.ready.is_set():
             raise HeadlessNotReady(NOT_READY_MESSAGE)
@@ -197,6 +201,128 @@ class HeadlessClientService(Service):
             for n in names:
                 status[n] += summary[n]
         return status
+
+    def exposed_find_user(self, username):
+        """
+        Search for the given username in all sessions, on all headless clients.
+        Returns a list of sessions they're currently connected to (NOT including
+        present worlds), and a list of sessions they're actually present in.
+        The search is case-insensitive.
+
+        The specific format for the returned value (a `dict`) is as follows:
+
+        {
+            "current": [(client_id, session_id)],
+            "present": [(client_id, session_id)]
+        }
+
+        The "current" and "present" keys are both a `list` of `tuple`s
+        consisting of both a `client_id` where the user was found, and the
+        `session_id` they are in. `session_id` is the same as the world number,
+        as shown by the `worlds` command in the headless client.
+
+        While Neos users cannot be present in more than one world at once, the
+        "present" `list` may contain more than one `tuple` if the user had just
+        switched sessions and the information in both sessions had not yet
+        updated to reflect this. It's up to the caller to decide what to do with
+        this information.
+
+        If the user is not connected to any headless clients at all, both of
+        these `list`s should be empty.
+        """
+        response = {"current": [], "present": []}
+        for cid in self.clients:
+            # Skip clients that are still starting up.
+            if not self.clients[cid].ready.is_set():
+                continue
+            for sid, w in enumerate(self.clients[cid].worlds()):
+                for u in self.clients[cid].users(sid):
+                    if u["name"].lower() == username.lower():
+                        if u["present"]:
+                            response["present"].append((cid, sid))
+                        else:
+                            response["current"].append((cid, sid))
+                        break
+        return response
+
+    def exposed_kick_from_all(self, username):
+        """
+        Attempt to kick the given username from all sessions, on all headless
+        clients. If the user exists in a session, they will be kicked. If not,
+        the command fails silently. Returns a `list` of clients and sessions
+        they were kicked from, in the format of `[(client_id, session_id)]`.
+        """
+
+        # Kicks are performed asynchronously on all headless clients.
+
+        asyncs, kicks = [], []
+        for cid in self.clients:
+            # Skip clients that are still starting up.
+            if not self.clients[cid].ready.is_set():
+                continue
+            for sid, w in enumerate(self.clients[cid].worlds()):
+                asyncs.append((cid, sid, self.clients[cid].async_(
+                    self.clients[cid].kick,
+                    username,
+                    world=sid
+                )))
+        for k in asyncs:
+            # TODO: Better exception handling?
+            if k[2].exception() == None:
+                kicks.append((k[0], k[1]))
+        return kicks
+
+    def exposed_ban_from_all(self, username, kick=False):
+        """
+        Attempt to ban the given username from all headless clients. Optionally,
+        also attempt to kick the user from all sessions. Bans apply at the
+        headless client level, meaning they restrict access to all sessions
+        hosted by a single client. Users cannot be banned per session.
+
+        Returns a `dict` with two keys: "bans", and "kicks". "bans" is a `list`
+        of client IDs, whereas "kicks" is a `list` of `tuple`s, containing both
+        client and session IDs in the format: `(client_id, session_id)`.
+
+        "bans" should almost always be a list of all headless clients currently
+        running, or an empty list. A filled list indicates the ban was
+        successful. An empty list indicates the username was not found and no
+        ban occurred, or there are no clients running.
+
+        "kicks" will only be populated if `kick` is set to `True`, and it
+        represents the clients and sessions that the user was kicked from, if
+        any. Otherwise, "kicks" will be an empty list.
+
+        `kick` is `False` by default because the `banByName` and `banByID`
+        commands in the headless client do not kick users by themselves. Setting
+        `kick` to `True` will perform a separate kick operations after the bans.
+        """
+
+        # Bans/kicks are performed asynchronously on all headless clients.
+
+        async_bans, async_kicks, bans, kicks = [], [], [], []
+        for cid in self.clients:
+            # Skip clients that are still starting up.
+            if not self.clients[cid].ready.is_set():
+                continue
+            async_bans.append((cid, self.clients[cid].async_(
+                    self.clients[cid].ban_by_name, username
+                )))
+            if kick:
+                for sid, w in enumerate(self.clients[cid].worlds()):
+                    async_kicks.append((cid, sid, self.clients[cid].async_(
+                        self.clients[cid].kick,
+                        username,
+                        world=sid
+                    )))
+        for b in async_bans:
+            # TODO: Better exception handling?
+            if b[1].exception() == None:
+                bans.append(b[0])
+        for k in async_kicks:
+            # TODO: Better exception handling?
+            if k[2].exception() == None:
+                kicks.append((k[0], k[1]))
+        return {"bans": bans, "kicks": kicks}
 
 if __name__ == "__main__":
     from rpyc.utils.server import ThreadedServer